+++ src/middleware/services/ollama/index.ts
@@ -1,2 +1,0 @@
-export * from './ollama.service';
-export * from './utils';